executable is now chash
