+++ output.txt
@@ -1,47 +1,43 @@
 Running 10 threads
-1721664571406607: WAITING ON INSERTS
-1721664571406629: INSERT,3449751099,Sid Meier,50000
-1721664571406651: WRITE LOCK ACQUIRED
-1721664571406608: INSERT,2500551319,Richard Garriot,40000
-1721664571406667: WRITE LOCK ACQUIRED
-1721664571406670: WRITE LOCK RELEASED
-1721664571406653: INSERT,1874280167,Shigeru Miyamoto,51000
-1721664571406729: WRITE LOCK ACQUIRED
-1721664571406735: WRITE LOCK RELEASED
-1721664571406744: INSERT,2569965317,Hideo Kojima,45000
-1721664571406766: WRITE LOCK ACQUIRED
-1721664571406661: WRITE LOCK RELEASED
-1721664571406772: WRITE LOCK RELEASED
-1721664571406825: INSERT,909366975,Roberta Williams,45900
-1721664571406834: WRITE LOCK ACQUIRED
-1721664571406837: WRITE LOCK RELEASED
-1721664571406848: INSERT,448054155,Gabe Newell,49000
-1721664571406870: WRITE LOCK ACQUIRED
-1721664571406875: DELETE AWAKENED
-1721664571406884: DELETE,Richard Garriot
-1721664571406913: WRITE LOCK ACQUIRED
-1721664571406918: WRITE LOCK RELEASED
-1721664571406877: WRITE LOCK RELEASED
-1721664571406859: DELETE AWAKENED
-1721664571406947: DELETE,Sid Meier
-1721664571406952: WRITE LOCK ACQUIRED
-1721664571406954: WRITE LOCK RELEASED
-1721664571406968: READ LOCK ACQUIRED
-1721664571406974: SEARCH: NOT FOUND NOT FOUND
-1721664571406977: READ LOCK RELEASED
-1721664571406986: INSERT,660427450,Carol Shaw,41000
-1721664571406995: WRITE LOCK ACQUIRED
-1721664571406998: WRITE LOCK RELEASED
-1721664571407102: READ LOCK ACQUIRED
-1721664571407114: SEARCH: NOT FOUND NOT FOUND
-1721664571407117: READ LOCK RELEASED
+1731614513338474: WRITE LOCK ACQUIRED
+1731614513338509: INSERT,2500551319,Richard Garriot,40000
+1731614513338522: WRITE LOCK RELEASED
+1731614513338530: WRITE LOCK ACQUIRED
+1731614513338532: INSERT,1874280167,Shigeru Miyamoto,51000
+1731614513338535: WRITE LOCK RELEASED
+1731614513338539: WRITE LOCK ACQUIRED
+1731614513338542: INSERT,3449751099,Sid Meier,50000
+1731614513338542: WRITE LOCK RELEASED
+1731614513338562: WAITING ON INSERTS
+1731614513338590: WRITE LOCK ACQUIRED
+1731614513338591: INSERT,2569965317,Hideo Kojima,45000
+1731614513338592: WRITE LOCK RELEASED
+1731614513338614: WRITE LOCK ACQUIRED
+1731614513338615: INSERT,448054155,Gabe Newell,49000
+1731614513338615: WRITE LOCK RELEASED
+1731614513338624: WRITE LOCK ACQUIRED
+1731614513338625: INSERT,909366975,Roberta Williams,45900
+1731614513338626: WRITE LOCK RELEASED
+1731614513338649: WAITING ON INSERTS
+1731614513338664: WRITE LOCK ACQUIRED
+1731614513338665: INSERT,660427450,Carol Shaw,41000
+1731614513338694: DELETE AWAKENED
+1731614513338695: WRITE LOCK ACQUIRED
+1731614513338695: DELETE,Sid Meier
+1731614513338700: WRITE LOCK RELEASED
+1731614513338702: DELETE AWAKENED
+1731614513338704: WRITE LOCK ACQUIRED
+1731614513338704: WRITE LOCK RELEASED
+1731614513338704: DELETE,Richard Garriot
+1731614513338716: WRITE LOCK RELEASED
+1731614513338720: READ LOCK ACQUIRED
+1731614513338720: SEARCH: NOT FOUND NOT FOUND
+1731614513338721: READ LOCK RELEASED
 Finished all threads.
-Number of lock acquisitions:  11
-Number of lock releases:  11
-1721664571407274: READ LOCK ACQUIRED
+Number of lock acquisitions: 10
+Number of lock releases: 10
 448054155,Gabe Newell,49000
 660427450,Carol Shaw,41000
 909366975,Roberta Williams,45900
 1874280167,Shigeru Miyamoto,51000
 2569965317,Hideo Kojima,45000
-1721664571407298: READ LOCK RELEASED
