--- output2.txt
@@ -1,43 +1,0 @@
-Running 10 threads
-1731556725382032: WRITE LOCK ACQUIRED
-1731556725382035: INSERT,2500551319,Richard Garriot,40000
-1731556725382051: WRITE LOCK RELEASED
-1731556725382056: WRITE LOCK ACQUIRED
-1731556725382156: WAITING ON INSERTS
-1731556725382171: INSERT,1874280167,Shigeru Miyamoto,51000
-1731556725382188: WRITE LOCK RELEASED
-1731556725382194: WRITE LOCK ACQUIRED
-1731556725382229: INSERT,3449751099,Sid Meier,50000
-1731556725382230: WRITE LOCK RELEASED
-1731556725382915: WRITE LOCK ACQUIRED
-1731556725382917: INSERT,2569965317,Hideo Kojima,45000
-1731556725382918: WRITE LOCK RELEASED
-1731556725382954: WRITE LOCK ACQUIRED
-1731556725382955: INSERT,448054155,Gabe Newell,49000
-1731556725382956: WRITE LOCK RELEASED
-1731556725383019: WRITE LOCK ACQUIRED
-1731556725383020: INSERT,909366975,Roberta Williams,45900
-1731556725383021: WRITE LOCK RELEASED
-1731556725383065: WAITING ON INSERTS
-1731556725383093: WRITE LOCK ACQUIRED
-1731556725383095: INSERT,660427450,Carol Shaw,41000
-1731556725383111: WRITE LOCK RELEASED
-1731556725383113: DELETE AWAKENED
-1731556725383114: WRITE LOCK ACQUIRED
-1731556725383115: DELETE,Sid Meier
-1731556725383117: DELETE AWAKENED
-1731556725383123: READ LOCK ACQUIRED
-1731556725383124: SEARCH: NOT FOUND NOT FOUND
-1731556725383124: READ LOCK RELEASED
-1731556725383133: WRITE LOCK RELEASED
-1731556725383137: WRITE LOCK ACQUIRED
-1731556725383137: DELETE,Richard Garriot
-1731556725383138: WRITE LOCK RELEASED
-Finished all threads.
-Number of lock acquisitions: 10
-Number of lock releases: 10
-448054155,Gabe Newell,49000
-660427450,Carol Shaw,41000
-909366975,Roberta Williams,45900
-1874280167,Shigeru Miyamoto,51000
-2569965317,Hideo Kojima,45000
